requirements.txt
google-ads==22.1.0
streamlit==1.31.0
python-dotenv==1.0.0
pandas==2.2.0
protobuf>=3.13.0,<5.0.0